geom_node_range {ggraph}	R Documentation

Show nodes as a line spanning a horizontal range

Description

This geom is most useful together with the fabric layout for showing the horizontal span of each node.

Usage

geom_node_range(mapping = NULL, data = NULL, position = "identity",
  show.legend = NA, ...)

Arguments

mapping
Set of aesthetic mappings created by ggplot2::aes() or ggplot2::aes_(). By default x is mapped to xmin, xend is mapped to xmax and y and yend are mapped to y in the node data.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Aesthetics

geom_node_point understand the following aesthetics. Bold aesthetics are automatically set, but can be overridden.

x

xend

y

yend

alpha

colour

linetype

size

filter

Author(s)

Thomas Lin Pedersen

See Also

Other geom_node_*: geom_node_arc_bar, geom_node_circle, geom_node_point, geom_node_text, geom_node_tile, geom_node_voronoi

Examples

require(tidygraph)
gr <- as_tbl_graph(highschool)

ggraph(gr, layout = 'fabric') +
  geom_node_range()

[Package ggraph version 2.0.0 Index]