ggraph-package {ggraph}	R Documentation

ggraph: An Implementation of Grammar of Graphics for Graphs and Networks

Description

The grammar of graphics as implemented in ggplot2 is a poor fit for graph and network visualizations due to its reliance on tabular data input. ggraph is an extension of the ggplot2 API tailored to graph visualizations and provides the same flexible approach to building up plots layer by layer.

Author(s)

Maintainer: Thomas Lin Pedersen thomasp85@gmail.com (0000-0002-5147-4711)

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://github.com/thomasp85/ggraph

Report bugs at https://github.com/thomasp85/ggraph/issues

[Package ggraph version 2.0.0 Index]