highschool {ggraph}	R Documentation

Friendship among high school boys

Description

This dataset shows the friendship among high school boys as assessed by the question: "What fellows here in school do you go around with most often?". The question was posed twice, with one year in between (1957 and 1958) and shows the evolution in friendship between the two timepoints.

Usage

highschool

Format

The graph is stored as an unnamed edgelist with a year attribute.

from

The boy answering the question

to

The boy being the answer to the question

year

The year the friendship was reported

Source

Coleman, J. S. Introduction to Mathematical Sociology. New York: Free Press, pp.450-451.

[Package ggraph version 2.0.0 Index]