layout_tbl_graph_backbone {ggraph}	R Documentation

Place node to emphasize group structure

Description

This layout is optimised for drawing small-world types of graphs often found in social networks, where distinct groups are still highly connected to the remaining graph. Typical layouts struggle with this as they attempt to minimise the edge length of all edges equally. The backbone layout is based on weighing edges based on how well they hold together communities. The end result is that communities tend to stick together despite high interconnectivity.

Usage

layout_tbl_graph_backbone(graph, keep = 0.2, circular = FALSE)

Arguments

graph
A tbl_graph object

keep
The fraction of edges to use for creating the backbone

circular
ignored

Value

A data.frame with the columns x, y, circular as well as any information stored as node variables in the tbl_graph object. Further an edge attribute called backbone is added giving whether the edge was selected as backbone.

Author(s)

The underlying algorithm is implemented in the graphlayouts package by David Schoch

References

Nocaj, A., Ortmann, M., & Brandes, U. (2015). Untangling the hairballs of multi-centered, small-world online social media networks. Journal of Graph Algorithms and Applications: JGAA, 19(2), 595-618.

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_centrality, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_eigen, layout_tbl_graph_fabric, layout_tbl_graph_focus, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_matrix, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap, layout_tbl_graph_unrooted

[Package ggraph version 2.0.0 Index]