layout_tbl_graph_centrality {ggraph}	R Documentation

Place nodes in circles according to centrality measure

Description

This layout places nodes in circles with the radii relative to a given centrality measure. Under the hood it use stress majorisation to place nodes optimally given the radius constraint.

Usage

layout_tbl_graph_centrality(graph, centrality, scale = TRUE,
  niter = 500, tolerance = 1e-04, tseq = seq(0, 1, 0.2),
  circular = FALSE)

Arguments

graph
A tbl_graph object

centrality
An expression evaluating to a centrality measure for the nodes. See the different centrality_*() algorithms in tidygraph for a selection.

scale
Should the centrality measure be scaled between 0 and 100

niter
number of iterations during stress optimization

tolerance
stopping criterion for stress optimization

tseq
Transitioning steps

circular
ignored

Value

A data.frame with the columns x, y, circular, centrality as well as any information stored as node variables in the tbl_graph object.

Author(s)

The underlying algorithm is implemented in the graphlayouts package by David Schoch

References

Brandes, U., & Pich, C. (2011). More flexible radial layout. Journal of Graph Algorithms and Applications, 15(1), 157-173.

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_backbone, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_eigen, layout_tbl_graph_fabric, layout_tbl_graph_focus, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_matrix, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap, layout_tbl_graph_unrooted

[Package ggraph version 2.0.0 Index]