layout_tbl_graph_eigen {ggraph}	R Documentation

Place nodes according to their eigenvalues

Description

This layout is based on the idea of spectral layouts where node coordinates are calculated directly by decomposing a matrix representation of the graph and extracting the eigenvectors.

Usage

layout_tbl_graph_eigen(graph, type = "laplacian",
  eigenvector = "smallest", circular = FALSE)

Arguments

graph
A tbl_graph object

type
The type of matrix to extract the eigenvectors from. Either 'laplacian' or 'adjacency'

eigenvector
The eigenvector to use for coordinates. Either 'smallest' or 'largest'

circular
ignored

Value

A data.frame with the columns x, y, circular as well as any information stored as node variables in the tbl_graph object.

Author(s)

The underlying algorithm is implemented in the graphlayouts package by David Schoch

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_backbone, layout_tbl_graph_centrality, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_fabric, layout_tbl_graph_focus, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_matrix, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap, layout_tbl_graph_unrooted

[Package ggraph version 2.0.0 Index]