layout_tbl_graph_fabric {ggraph}	R Documentation

Create a fabric layout

Description

This layout is a bit unusual in that it shows nodes as horizontal line ranges end edges as evenly spaced vertical spans connecting the nodes. As with the matrix layout the strength comes from better scalability but its use require some experience recognising the patterns that different connectivity features gives rise to. As with matrix layouts the ordering of nodes have huge power over the look of the plot. The node_rank_fabric() mimics the default ordering from the original BioFabric implementation, but other ranking algorithms from tidygraph can be used with the sort.by argument as well. Fabric layouts tend to become quite wide as the graph grows which is something that should be handled with care - e.g. by only zooming in on a specific region.

Usage

layout_tbl_graph_fabric(graph, circular = FALSE, sort.by = NULL,
  shadow.edges = FALSE)

node_rank_fabric()

Arguments

graph
An tbl_graph object

circular
Ignored

sort.by
An expression providing the sorting of the nodes. If NULL the nodes will be ordered by their index in the graph.

shadow.edges
Should shadow edges be shown.

Value

A data.frame with the columns x, xmin, xmax, y, circular as well as any information stored as node variables in the tbl_graph object. Further, the edges of the graph will gain a edge_x variable giving the horizontal position of the edge as well as a shadow_edge variable denoting whether the edge is a shadow edge added by the layout.

References

BioFabric website: http://www.biofabric.org

Longabaugh, William J.R. (2012). Combing the hairball with BioFabric: a new approach for visualization of large networks. BMC Bioinformatics, 13: 275. https://doi.org/10.1186/1471-2105-13-275

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_backbone, layout_tbl_graph_centrality, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_eigen, layout_tbl_graph_focus, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_matrix, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap, layout_tbl_graph_unrooted

[Package ggraph version 2.0.0 Index]