layout_tbl_graph_focus {ggraph}	R Documentation

Place nodes in circles based on distance to a specific node

Description

This layout constrains node placement to a radius relative to its distance to a given node. It then uses stress majorisation to find an optimal node distribution according to this constraint.

Usage

layout_tbl_graph_focus(graph, focus, weights = NULL, niter = 500,
  tolerance = 1e-04, circular = TRUE)

Arguments

graph
a tbl_graph object

focus
An expression evaluating to a selected node. Can either be a single integer or a logical vector with a single TRUE element.

weights
An expression evaluated on the edge data to provide edge weights for the layout. Currently ignored for the sparse version

niter
number of iterations during stress optimization

tolerance
stopping criterion for stress optimization

circular
ignored

Value

A data.frame with the columns x, y, circular, distance as well as any information stored as node variables in the tbl_graph object.

Author(s)

The underlying algorithm is implemented in the graphlayouts package by David Schoch

References

Brandes, U., & Pich, C. (2011). More flexible radial layout. Journal of Graph Algorithms and Applications, 15(1), 157-173.

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_backbone, layout_tbl_graph_centrality, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_eigen, layout_tbl_graph_fabric, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_matrix, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap, layout_tbl_graph_unrooted

[Package ggraph version 2.0.0 Index]