layout_tbl_graph_linear {ggraph}	R Documentation

Place nodes on a line or circle

Description

This layout puts all nodes on a line, possibly sorted by a node attribute. If circular = TRUE the nodes will be laid out on the unit circle instead. In the case where the sort.by attribute is numeric, the numeric values will be used as the x-position and it is thus possible to have uneven spacing between the nodes.

Usage

layout_tbl_graph_linear(
  graph,
  circular,
  sort.by = NULL,
  use.numeric = FALSE,
  offset = pi/2,
  weight = NULL
)

Arguments

graph
An tbl_graph object

circular
Logical. Should the layout be transformed to a circular representation. Defaults to FALSE.

sort.by
The name of a node variable to sort the nodes by.

use.numeric
Logical. Should a numeric sort.by attribute be used as the actual x-coordinates in the layout. May lead to overlapping nodes. Defaults to FALSE

offset
If circular = TRUE, where should it begin. Defaults to pi/2 which is equivalent to 12 o'clock.

weight
A weight for each node. Nodes will be spread out according to their weight so that nodes with heigher weight will have more space around them. Ignored if use.numeric = TRUE

Value

A data.frame with the columns x, y, circular as well as any information stored as node variables in the tbl_graph object. Further, if circular = FALSE a width column and if circular = TRUE a start, end, and r0 column.

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto(), layout_tbl_graph_backbone(), layout_tbl_graph_cactustree(), layout_tbl_graph_centrality(), layout_tbl_graph_circlepack(), layout_tbl_graph_dendrogram(), layout_tbl_graph_eigen(), layout_tbl_graph_fabric(), layout_tbl_graph_focus(), layout_tbl_graph_hive(), layout_tbl_graph_htree(), layout_tbl_graph_igraph(), layout_tbl_graph_manual(), layout_tbl_graph_matrix(), layout_tbl_graph_metro(), layout_tbl_graph_partition(), layout_tbl_graph_pmds(), layout_tbl_graph_sf(), layout_tbl_graph_stress(), layout_tbl_graph_treemap(), layout_tbl_graph_unrooted()

[Package ggraph version 2.2.2 Index]