layout_tbl_graph_matrix {ggraph}	R Documentation

Place nodes on a diagonal

Description

This layout puts all nodes on a diagonal, thus preparing the layout for use with geom_edge_point() resulting in a matrix layout. While matrix layouts excel in scalability, the interpretation of the visual is very dependent on the sorting of the nodes. Different sorting algorithms have been implemented in tidygraph and these can be used directly. Behrisch et al. (2016) have provided a nice overview of some of the different sorting algorithms and what insight they might bring, along with a rundown of different patterns to look out for.

Usage

layout_tbl_graph_matrix(graph, circular = FALSE, sort.by = NULL)

Arguments

graph
An tbl_graph object

circular
Ignored

sort.by
An expression providing the sorting of the nodes. If NULL the nodes will be ordered by their index in the graph.

Value

A data.frame with the columns x, y, circular as well as any information stored as node variables in the tbl_graph object.

References

Behrisch, M., Bach, B., Riche, N. H., Schreck, T., Fekete, J.-D. (2016). Matrix Reordering Methods for Table and Network Visualization. Computer Graphics Forum, 35: 693–716. http://doi.org/10.1111/cgf.12935

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_backbone, layout_tbl_graph_centrality, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_eigen, layout_tbl_graph_fabric, layout_tbl_graph_focus, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap, layout_tbl_graph_unrooted

[Package ggraph version 2.0.0 Index]