layout_tbl_graph_sf {ggraph}	R Documentation

Place nodes on their geographical space

Description

This layout is built for objects of class sfnetwork and is meant to plot a graph on its geographical space, by extracting its X and Y coordinates

Usage

layout_tbl_graph_sf(graph, circular = FALSE)

Arguments

graph
An sfnetwork object

circular
ignored

Value

A data.frame with the columns x, y, circular.

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto(), layout_tbl_graph_backbone(), layout_tbl_graph_cactustree(), layout_tbl_graph_centrality(), layout_tbl_graph_circlepack(), layout_tbl_graph_dendrogram(), layout_tbl_graph_eigen(), layout_tbl_graph_fabric(), layout_tbl_graph_focus(), layout_tbl_graph_hive(), layout_tbl_graph_htree(), layout_tbl_graph_igraph(), layout_tbl_graph_linear(), layout_tbl_graph_manual(), layout_tbl_graph_matrix(), layout_tbl_graph_metro(), layout_tbl_graph_partition(), layout_tbl_graph_pmds(), layout_tbl_graph_stress(), layout_tbl_graph_treemap(), layout_tbl_graph_unrooted()

[Package ggraph version 2.2.2 Index]