layout_tbl_graph_unrooted {ggraph}	R Documentation

Create an unrooted layout using equal-angle or equal-daylight

Description

When drawing unrooted trees the standard dendrogram layout is a bad fit as it implicetly creates a visual root node. Instead it is possible to spread the leafs out on the plane without putting any special emphasis on a particular node using an unrooted layout. The standard algorithm is the equal angle algorithm, but it can struggle with optimising the leaf distribution for large trees trees with very uneven branch lenght. The equal daylight algorithm modifies the output of the equal angle algorithm to better disperse the leaves, at the cost of higher computational cost and the possibility of edge crossings for very large unbalanced trees. For standard sized trees the daylight algorithm is far superior and not too heavy so it is the default.

Usage

layout_tbl_graph_unrooted(graph, daylight = TRUE, length = NULL,
  tolerance = 0.05, rotation_mod = 1, maxiter = 100,
  circular = FALSE)

Arguments

graph
A tbl_graph object

daylight
Should equal-daylight adjustments be made

length
An expression evaluating to the branch length of each edge

tolerance
The threshold for mean angular adjustment before terminating the daylight adjustment

rotation_mod
A modifier for the angular adjustment of each branch. Set it below 1 to let the daylight adjustment progress more slowly

maxiter
The maximum number of iterations in the the daylight adjustment

circular
ignored

Value

A data.frame with the columns x, y, circular, leaf as well as any information stored as node variables in the tbl_graph object.

Note

Unrooted is a layout intended for undirected trees, that is, graphs with no cycles. If the provided graph does not fit this format an attempt to convert it to such a format will be made.

References

Felsenstein, J. (2004) Drawing Trees, in Inferring Phylogenies. Sinauer Assoc., pp 573-584

See Also

Other layout_tbl_graph_*: layout_tbl_graph_auto, layout_tbl_graph_backbone, layout_tbl_graph_centrality, layout_tbl_graph_circlepack, layout_tbl_graph_dendrogram, layout_tbl_graph_eigen, layout_tbl_graph_fabric, layout_tbl_graph_focus, layout_tbl_graph_hive, layout_tbl_graph_igraph, layout_tbl_graph_linear, layout_tbl_graph_manual, layout_tbl_graph_matrix, layout_tbl_graph_partition, layout_tbl_graph_pmds, layout_tbl_graph_stress, layout_tbl_graph_treemap

[Package ggraph version 2.0.0 Index]