layout_to_table {ggraph}	R Documentation

Convert a layout to a table

Description

This generic takes care of dispatching various layout types (names, functions, tables) to their respective functions that will return a valid layout table.

Usage

layout_to_table(layout, graph, ...)

Arguments

layout
A supported object

graph
A tbl_graph

...
passed on to implementations

Value

A valid data.frame

[Package ggraph version 2.0.0 Index]