makeContent.cappedpathgrob {ggraph}	R Documentation

Dynamic capping of paths

Description

This function takes care of updating which parts of the paths get removed when the device dimensions are updated.

Usage

## S3 method for class 'cappedpathgrob'
makeContent(x)

[Package ggraph version 2.0.0 Index]