scale_edge_alpha {ggraph}	R Documentation

Edge alpha scales

Description

This set of scales defines new alpha scales for edge geoms equivalent to the ones already defined by ggplot2. See ggplot2::scale_alpha() for more information. The different geoms will know whether to use edge scales or the standard scales so it is not necessary to write edge_alpha in the call to the geom - just use alpha.

Usage

scale_edge_alpha(..., range = c(0.1, 1))

scale_edge_alpha_continuous(..., range = c(0.1, 1))

scale_edge_alpha_discrete(..., range = c(0.1, 1))

scale_edge_alpha_binned(..., range = c(0.1, 1))

scale_edge_alpha_manual(..., values, breaks = waiver(), na.value = NA)

scale_edge_alpha_identity(..., guide = "none")

Arguments

...
Other arguments passed on to continuous_scale(), binned_scale(), or discrete_scale() as appropriate, to control name, limits, breaks, labels and so forth.

range
Output range of alpha values. Must lie between 0 and 1.

values
a set of aesthetic values to map data values to. The values will be matched in order (usually alphabetical) with the limits of the scale, or with breaks if provided. If this is a named vector, then the values will be matched based on the names instead. Data values that don't match will be given na.value.

breaks
One of:

NULL for no breaks

waiver() for the default breaks (the scale limits)

A character vector of breaks

A function that takes the limits as input and returns breaks as output

na.value
The aesthetic value to use for missing (NA) values

guide
Guide to use for this scale. Defaults to "none".

Value

A ggproto object inheriting from Scale

See Also

Other scale_edge_*: scale_edge_colour, scale_edge_fill, scale_edge_linetype(), scale_edge_shape(), scale_edge_size(), scale_edge_width(), scale_label_size()

[Package ggraph version 2.2.2 Index]