scale_type.geometry {ggraph}	R Documentation

Define default scale type for geometry

Description

This function is quite useless as geometry is not meant to be scaled, but it is a requirement for ggplot2 to handle it correctly.

Usage

## S3 method for class 'geometry'
scale_type(x)

[Package ggraph version 2.0.0 Index]