Apply a function recursively to a tree

This function allows for easy recursive function calling of tree-like structures. The recursion can happen either downwards, calling the children after the parent, or upwards, calling the parent after the children. At each recursion there is access to the node(s) that was/were called before and thus have already been through the function.

tree_apply(tree, FUN, ...)

# S3 method for default
tree_apply(tree, FUN, ...)

# S3 method for igraph
tree_apply(tree, FUN, direction = "down", mode = "out",
  ...)

# S3 method for dendrogram
tree_apply(tree, FUN, direction = "down", ...)

Arguments

tree
A tree-like object. Currently support for igraph objects

FUN
The function to apply to each node. The function must return a modified version of the graph if class(tree) == 'igraph' or a modified version of the node if class(tree) == 'dendrogram'

...
Additional parameters to FUN

direction
The direction of the recursion. If direction = 'down' The parent will get handled before the children, while the reverse is true if direction = 'up'

mode
For class(tree) == 'igraph' the directionality of the edges in the graph. If mode = 'out' then parents points towards their children, while the reverse is true for mode = 'in'

Value

A modified version of tree (if anything is modified in FUN)

Details

The function is called with a set of predefined parameters along with user defined ones. If direction = 'down' The parameters supplied automatically to the function are: node, parent, depth and tree, while if direction = 'up' the parameters are: node, children, depth and tree. The nature of node, parent and children depends on the class of tree. If class(tree) == 'igraph' they will be indices of the relevant vertices in the graph. If class(tree) == 'dendrogram' they will be the actual dendrogram objects.