get_nodes {ggraph}	R Documentation

Create a node extractor function

Description

This function returns another function that can extract nodes from a ggraph_layout object. As a ggraph_layout object is essentially a data.frame of nodes it might seem useless to provide this function, but since the node data is not necessarily available until after the ggraph() call it can be beneficial to be able to add information to the node data on a per-layer basis. Unlike get_edges() the use of get_nodes is not mandatory and is only required if additional data should be added to selected node layers.

Usage

get_nodes(...)

Arguments

...
Additional data that should be cbind'ed together with the node data.

Value

A data.frame with the node data as well of any additional data supplied through ...

See Also

Other extractors: get_con, get_edges

[Package ggraph version 2.0.0 Index]