Place nodes in circles according to centrality measure

This layout places nodes in circles with the radii relative to a given centrality measure. Under the hood it use stress majorisation to place nodes optimally given the radius constraint.

layout_tbl_graph_centrality(
  graph,
  centrality,
  scale = TRUE,
  niter = 500,
  tolerance = 1e-04,
  tseq = seq(0, 1, 0.2),
  circular = FALSE
)

Arguments

graph

A tbl_graph object

centrality

An expression evaluating to a centrality measure for the nodes. See the different centrality_*() algorithms in tidygraph for a selection.

scale

Should the centrality measure be scaled between 0 and 100

niter

number of iterations during stress optimization

tolerance

stopping criterion for stress optimization

tseq

Transitioning steps

circular

ignored

Value

A data.frame with the columns x, y, circular, centrality as well as any information stored as node variables in the tbl_graph object.

References

Brandes, U., & Pich, C. (2011). More flexible radial layout. Journal of Graph Algorithms and Applications, 15(1), 157-173.

See also

Author

The underlying algorithm is implemented in the graphlayouts package by David Schoch