Place nodes according to their eigenvalues

This layout is based on the idea of spectral layouts where node coordinates are calculated directly by decomposing a matrix representation of the graph and extracting the eigenvectors.

layout_tbl_graph_eigen(
  graph,
  type = "laplacian",
  eigenvector = "smallest",
  circular = FALSE
)

Arguments

graph
A tbl_graph object

type
The type of matrix to extract the eigenvectors from. Either 'laplacian' or 'adjacency'

eigenvector
The eigenvector to use for coordinates. Either 'smallest' or 'largest'

circular
ignored

Value

A data.frame with the columns x, y, circular as well as any information stored as node variables in the tbl_graph object.

See also