scale_edge_alpha {ggraph}	R Documentation

Edge alpha scales

Description

This set of scales defines new alpha scales for edge geoms equivalent to the ones already defined by ggplot2. See ggplot2::scale_alpha() for more information. The different geoms will know whether to use edge scales or the standard scales so it is not necessary to write edge_alpha in the call to the geom - just use alpha.

Usage

scale_edge_alpha(..., range = c(0.1, 1))

scale_edge_alpha_continuous(..., range = c(0.1, 1))

scale_edge_alpha_discrete(..., range = c(0.1, 1))

scale_edge_alpha_manual(..., values)

scale_edge_alpha_identity(..., guide = "none")

Arguments

...
Other arguments passed on to continuous_scale() or discrete_scale() as appropriate, to control name, limits, breaks, labels and so forth.

range
Output range of alpha values. Must lie between 0 and 1.

values
a set of aesthetic values to map data values to. If this is a named vector, then the values will be matched based on the names. If unnamed, values will be matched in order (usually alphabetical) with the limits of the scale. Any data values that don't match will be given na.value.

guide
Guide to use for this scale. Defaults to "none".

Value

A ggproto object inheriting from Scale

See Also

Other scale_edge_*: scale_edge_colour, scale_edge_fill, scale_edge_linetype, scale_edge_shape, scale_edge_size, scale_edge_width, scale_label_size

[Package ggraph version 2.0.0 Index]